requirements.txt
# Core AI and LLM libraries
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0

# Web interface
gradio<6.0

# pdf processing
pypdf>=5.0.0

# General utilities
python-dotenv>=1.0.0
